feature/favorites

+++ src/pages/About.vue
@@ -14,20 +14,35 @@ const favorites = ref([])
 
 const deleteFavorite = async (item) => {
   try {
+    if (!item.favoriteId) {
+      throw new Error('favoriteId is null')
+    }
     await axios.delete(`https://3afc0b251db73ab9.mokky.dev/favorites/${item.favoriteId}`)
-    favorites.value = favorites.value.filter(favorite => favorite.item_id !== item.id);
+    favorites.value = favorites.value.filter(favorite => favorite.favoriteId !== item.favoriteId)
+    item.isFavorite = false
+    item.favoriteId = null
   } catch (err) {
     console.log(err)
   }
 }
 
-onMounted(async () => {
+const fetchFavorites = async () => {
   try {
-      const { data } = await axios.get('https://3afc0b251db73ab9.mokky.dev/favorites')
-      favorites.value = data.map((obj) => obj.item)
+    const { data } = await axios.get('https://3afc0b251db73ab9.mokky.dev/favorites')
+    favorites.value = data.map((obj) => {
+      const item = obj.item
+      item.favoriteId = obj.id // Присваиваем favoriteId для каждого элемента
+      return item
+    })
+    console.log(favorites.value) // Для проверки корректности данных
   } catch (err) {
     console.log(err)
   }
+}
+
+
+onMounted(() => {
+  fetchFavorites()
 })
 
 </script>
